refactor: remove deprecated curator ServiceAnnouncer stuff
