--- server/src/main/java/org/apache/druid/curator/discovery/CuratorServiceUtils.java
@@ -1,43 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.druid.curator.discovery;
-
-/**
- * This class is only used by Deprecated classes.
- */
-@Deprecated
-public class CuratorServiceUtils
-{
-  /**
-   * Replacing '/' with ':' in service names makes it easier to provide an HTTP interface using
-   * <a href="http://curator.apache.org/curator-x-discovery-server/">curator-x-discovery-server</a>.
-   *
-   * <p>This method is marked protected because it should never be used outside of the
-   * {@code org.apache.druid.curator.discovery} package. If you are tempted to use this method anywhere else you are
-   * most likely doing something wrong. Mapping the actual service name to the name used within curator should be left
-   * to {@link CuratorServiceAnnouncer}.
-   *
-   * @see CuratorServiceAnnouncer
-   */
-  protected static String makeCanonicalServiceName(String serviceName)
-  {
-    return serviceName.replace('/', ':');
-  }
-}
