+++ server/src/main/java/org/apache/druid/segment/realtime/ChatHandler.java
@@ -20,7 +20,7 @@
 package org.apache.druid.segment.realtime;
 
 /**
- * Objects that can be registered with a {@link ServiceAnnouncingChatHandlerProvider} and provide http endpoints for indexing-related
+ * Objects that can be registered with a {@link LocalChatHandlerProvider} and provide http endpoints for indexing-related
  * objects. This interface is empty because it only exists to signal intent. The actual http endpoints are provided
  * through JAX-RS annotations on the {@link ChatHandler} objects.
  */
