+++ server/src/main/java/org/apache/druid/segment/realtime/NoopChatHandlerProvider.java
@@ -32,12 +32,6 @@ public class NoopChatHandlerProvider implements ChatHandlerProvider
   }
 
   @Override
-  public void register(String key, ChatHandler handler, boolean announce)
-  {
-    // do nothing
-  }
-
-  @Override
   public void unregister(String key)
   {
     // do nothing
