+++ server/src/test/java/org/apache/druid/segment/realtime/LocalChatHandlerProviderTest.java
@@ -1,0 +1,64 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.druid.segment.realtime;
+
+import org.apache.druid.java.util.common.ISE;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+
+public class LocalChatHandlerProviderTest
+{
+  private static class TestChatHandler implements ChatHandler
+  {
+  }
+
+  private static final String TEST_SERVICE_NAME = "test-service-name";
+
+  private LocalChatHandlerProvider chatHandlerProvider;
+
+  @Before
+  public void setUp()
+  {
+    chatHandlerProvider = new LocalChatHandlerProvider();
+  }
+
+  @Test
+  public void testRegisterAndGet()
+  {
+    ChatHandler testChatHandler = new TestChatHandler();
+
+    Assert.assertFalse("bad initial state", chatHandlerProvider.get(TEST_SERVICE_NAME).isPresent());
+
+    chatHandlerProvider.register(TEST_SERVICE_NAME, testChatHandler);
+    Assert.assertTrue("chatHandler did not register", chatHandlerProvider.get(TEST_SERVICE_NAME).isPresent());
+    Assert.assertEquals(testChatHandler, chatHandlerProvider.get(TEST_SERVICE_NAME).get());
+
+    chatHandlerProvider.unregister(TEST_SERVICE_NAME);
+    Assert.assertFalse("chatHandler did not deregister", chatHandlerProvider.get(TEST_SERVICE_NAME).isPresent());
+  }
+
+  @Test
+  public void testDuplicateRegistrationThrows()
+  {
+    chatHandlerProvider.register(TEST_SERVICE_NAME, new TestChatHandler());
+    Assert.assertThrows(ISE.class, () -> chatHandlerProvider.register(TEST_SERVICE_NAME, new TestChatHandler()));
+  }
+}
