+++ server/src/test/java/org/apache/druid/server/coordinator/DruidCoordinatorTest.java
@@ -31,7 +31,6 @@ import org.apache.druid.client.ImmutableDruidDataSource;
 import org.apache.druid.client.ImmutableDruidServer;
 import org.apache.druid.client.ServerInventoryView;
 import org.apache.druid.common.config.JacksonConfigManager;
-import org.apache.druid.curator.discovery.LatchableServiceAnnouncer;
 import org.apache.druid.discovery.DruidLeaderSelector;
 import org.apache.druid.java.util.common.Intervals;
 import org.apache.druid.java.util.common.concurrent.ScheduledExecutorFactory;
@@ -103,8 +102,6 @@ public class DruidCoordinatorTest
   private ScheduledExecutorFactory scheduledExecutorFactory;
   private LoadQueueTaskMaster loadQueueTaskMaster;
   private MetadataRuleManager metadataRuleManager;
-  private CountDownLatch leaderAnnouncerLatch;
-  private CountDownLatch leaderUnannouncerLatch;
   private DruidCoordinatorConfig druidCoordinatorConfig;
   private DruidNode druidNode;
   private OverlordClient overlordClient;
@@ -147,8 +144,6 @@ public class DruidCoordinatorTest
     );
     druidNode = new DruidNode("hey", "what", false, 1234, null, true, false);
     scheduledExecutorFactory = ScheduledExecutors::fixed;
-    leaderAnnouncerLatch = new CountDownLatch(1);
-    leaderUnannouncerLatch = new CountDownLatch(1);
     serviceEmitter = new LatchableServiceEmitter();
     coordinator = new DruidCoordinator(
         druidCoordinatorConfig,
@@ -159,8 +154,6 @@ public class DruidCoordinatorTest
         overlordClient,
         loadQueueTaskMaster,
         new SegmentLoadQueueManager(serverInventoryView, loadQueueTaskMaster),
-        new LatchableServiceAnnouncer(leaderAnnouncerLatch, leaderUnannouncerLatch),
-        druidNode,
         new CoordinatorCustomDutyGroups(ImmutableSet.of()),
         EasyMock.createNiceMock(LookupCoordinatorManager.class),
         new TestDruidLeaderSelector(),
@@ -242,7 +235,6 @@ public class DruidCoordinatorTest
     Assert.assertNull(coordinator.getBroadcastSegments());
 
     // Wait for this coordinator to become leader
-    leaderAnnouncerLatch.await();
 
     // This coordinator should be leader by now
     Assert.assertTrue(coordinator.isLeader());
@@ -291,7 +283,6 @@ public class DruidCoordinatorTest
     Assert.assertEquals(Integer.valueOf(2), coordinator.getReplicationFactor(dataSegment.getId()));
 
     coordinator.stop();
-    leaderUnannouncerLatch.await();
 
     Assert.assertFalse(coordinator.isLeader());
     Assert.assertNull(coordinator.getCurrentLeader());
@@ -338,7 +329,6 @@ public class DruidCoordinatorTest
     EasyMock.replay(metadataRuleManager, serverInventoryView, loadQueueTaskMaster);
 
     coordinator.start();
-    leaderAnnouncerLatch.await(); // Wait for this coordinator to become leader
 
     serviceEmitter.coordinatorRunLatch.await();
 
@@ -359,7 +349,6 @@ public class DruidCoordinatorTest
     dataSegments.forEach(dataSegment -> Assert.assertEquals(Integer.valueOf(1), coordinator.getReplicationFactor(dataSegment.getId())));
 
     coordinator.stop();
-    leaderUnannouncerLatch.await();
 
     EasyMock.verify(serverInventoryView);
     EasyMock.verify(segmentsMetadataManager);
@@ -422,7 +411,6 @@ public class DruidCoordinatorTest
     EasyMock.replay(metadataRuleManager, serverInventoryView, loadQueueTaskMaster);
 
     coordinator.start();
-    leaderAnnouncerLatch.await(); // Wait for this coordinator to become leader
 
     serviceEmitter.coordinatorRunLatch.await();
 
@@ -447,7 +435,6 @@ public class DruidCoordinatorTest
     Assert.assertEquals(0L, underReplicationCountsPerDataSourcePerTierUsingClusterView.get(tierName2).getLong(dataSource));
 
     coordinator.stop();
-    leaderUnannouncerLatch.await();
 
     EasyMock.verify(serverInventoryView);
     EasyMock.verify(segmentsMetadataManager);
@@ -472,8 +459,6 @@ public class DruidCoordinatorTest
         overlordClient,
         loadQueueTaskMaster,
         null,
-        new LatchableServiceAnnouncer(leaderAnnouncerLatch, leaderUnannouncerLatch),
-        druidNode,
         emptyCustomDutyGroups,
         EasyMock.createNiceMock(LookupCoordinatorManager.class),
         new TestDruidLeaderSelector(),
@@ -525,8 +510,6 @@ public class DruidCoordinatorTest
         overlordClient,
         loadQueueTaskMaster,
         null,
-        new LatchableServiceAnnouncer(leaderAnnouncerLatch, leaderUnannouncerLatch),
-        druidNode,
         customDutyGroups,
         EasyMock.createNiceMock(LookupCoordinatorManager.class),
         new TestDruidLeaderSelector(),
@@ -578,8 +561,6 @@ public class DruidCoordinatorTest
         overlordClient,
         loadQueueTaskMaster,
         null,
-        new LatchableServiceAnnouncer(leaderAnnouncerLatch, leaderUnannouncerLatch),
-        druidNode,
         customDutyGroups,
         EasyMock.createNiceMock(LookupCoordinatorManager.class),
         new TestDruidLeaderSelector(),
@@ -689,8 +670,6 @@ public class DruidCoordinatorTest
         overlordClient,
         loadQueueTaskMaster,
         new SegmentLoadQueueManager(serverInventoryView, loadQueueTaskMaster),
-        new LatchableServiceAnnouncer(leaderAnnouncerLatch, leaderUnannouncerLatch),
-        druidNode,
         groups,
         EasyMock.createNiceMock(LookupCoordinatorManager.class),
         new TestDruidLeaderSelector(),
@@ -784,7 +763,6 @@ public class DruidCoordinatorTest
     coordinator.start();
     
     // Wait for this coordinator to become leader
-    leaderAnnouncerLatch.await();
 
     // This coordinator should be leader by now
     Assert.assertTrue(coordinator.isLeader());
@@ -817,7 +795,6 @@ public class DruidCoordinatorTest
     Assert.assertEquals(1, numsDeepStorageOnlySegmentsPerDataSource.getInt(dataSource));
 
     coordinator.stop();
-    leaderUnannouncerLatch.await();
 
     Assert.assertFalse(coordinator.isLeader());
     Assert.assertNull(coordinator.getCurrentLeader());
