+++ services/src/main/java/org/apache/druid/cli/CliPeon.java
@@ -113,8 +113,8 @@ import org.apache.druid.segment.loading.OmniDataSegmentKiller;
 import org.apache.druid.segment.loading.OmniDataSegmentMover;
 import org.apache.druid.segment.loading.StorageLocation;
 import org.apache.druid.segment.realtime.ChatHandlerProvider;
+import org.apache.druid.segment.realtime.LocalChatHandlerProvider;
 import org.apache.druid.segment.realtime.NoopChatHandlerProvider;
-import org.apache.druid.segment.realtime.ServiceAnnouncingChatHandlerProvider;
 import org.apache.druid.segment.realtime.appenderator.AppenderatorsManager;
 import org.apache.druid.segment.realtime.appenderator.PeonAppenderatorsManager;
 import org.apache.druid.server.DruidNode;
@@ -423,19 +423,19 @@ public class CliPeon extends GuiceRunnable
         binder,
         "druid.indexer.task.chathandler.type",
         Key.get(ChatHandlerProvider.class),
-        Key.get(ServiceAnnouncingChatHandlerProvider.class)
+        Key.get(LocalChatHandlerProvider.class)
     );
     final MapBinder<String, ChatHandlerProvider> handlerProviderBinder =
         PolyBind.optionBinder(binder, Key.get(ChatHandlerProvider.class));
     handlerProviderBinder
-        .addBinding("announce")
-        .to(ServiceAnnouncingChatHandlerProvider.class)
+        .addBinding("local")
+        .to(LocalChatHandlerProvider.class)
         .in(LazySingleton.class);
     handlerProviderBinder
         .addBinding("noop")
         .to(NoopChatHandlerProvider.class)
         .in(LazySingleton.class);
-    binder.bind(ServiceAnnouncingChatHandlerProvider.class).in(LazySingleton.class);
+    binder.bind(LocalChatHandlerProvider.class).in(LazySingleton.class);
     binder.bind(NoopChatHandlerProvider.class).in(LazySingleton.class);
   }
